feat: OGP 取得 REST API を追加 (T111)
OGP メタデータ取得ロジックを lib に切り出し、UI 用 Server Action と外部 API で
共有する。

- lib/ogp.ts: fetchOgpData(url) を新設（SSRF 対策・charset 判定・OGP パース、認証なし）
- fetchOgp.ts（Server Action）: getSession→redirect の後 fetchOgpData に委譲する薄いラッパに
- GET /api/ogp?url=（API キー認証・200 {title,image} / 取得失敗・SSRF は 200 空返し）
- proxy.ts の public ルートに /api/ogp(.*) を追加
- OGP ロジックのテストを lib/ogp.test.ts に移設＋route/Server Action テスト新設
- docs/api.md・docs/auth.md 更新
