+++ src/proxy.ts
@@ -8,6 +8,7 @@ const isPublicRoute = createRouteMatcher([
   // 外部連携用 REST API は Clerk ではなく API キー認証で保護する
   "/api/bookmarks(.*)",
   "/api/tags(.*)",
+  "/api/ogp(.*)",
 ]);
 
 export default clerkMiddleware(async (auth, request) => {
